fix auto update
